Transplanted door code
Copied to logic airlocks use to decide whether or not to open into the
windoor, since they are effectively the same. An issue was reported
where Perseus enforcers could access areas they should not be able to if
the area was restricted by a windoor with the same access as a regular
airlock. Tested and confirmed this code worked. Too tired to figure out
what was wrong with the other code.