Fixed airlocks not closing correctly with hacked timing
The issue was that the autoclose() proc was being called too quickly to
actually close the door

additionally i setup a if statement in the autoclose proc (Where it
checks if a human is there) that if the timing is hacked it would close
faster. tested and people walking through doors can still make it
through before the doors close quickly. tested while running too, works
fine.