Changed Chembooks to Generate Once
Modified the chem book code to generate books into html/manuals/ if there is not an entry for that book type.
This allows me to generate the book in advance, vastly reducing the load on the server.
If anyone needs to change the guide, they just delete the file and test the server. The books will be re-generated at that time.