Brains no longer take eye damage
Fix for #237

Brains added to the list of /mob/living/carbon types that cannot be
eyestabbed, along with aliens and slimes.

Brains can still take brute damage from being attacked by objects, as it
appears to be a feature according to the comment "//Oh noooeeeee" in the
MMI.dm file.